Refactor even more
